--- components/eventForm/index.ts
@@ -1,4 +1,0 @@
-export * from "./EventForm"
-export * from "./EventFormFields"
-export * from "./EventFormActionButtons"
-export * from "./toolbar"
